Add test case and fix imports

+++ include/gauxc/shell.hpp
@@ -14,6 +14,8 @@
 #include <array>
 #include <cmath>
 #include <iostream>
+#include <sstream>
+#include <stdexcept>
 #include <cassert>
 #include <algorithm>
 #include <tuple>
@@ -138,7 +140,7 @@ public:
     alpha_( alpha ), coeff_( coeff ), O_( O ),
     nprim_( nprim.get() ), l_( l.get() ), pure_( pure.get() ) {
 
-    if(nprim_ > detail::shell_nprim_max) {
+    if(nprim_ > static_cast<int32_t>(detail::shell_nprim_max)) {
       std::ostringstream oss;
       oss << "Shell has too many primitives! expected <= " << detail::shell_nprim_max << " actual value " << nprim_ << "!\n";
       throw std::out_of_range(oss.str());
